feat: replace old access form page with new one
