+++ frontend/src/router/index.js
@@ -2,7 +2,6 @@ import { createRouter, createWebHistory } from 'vue-router'
 import OidcCallback from '@/components/OidcCallback.vue'
 import HomePage from '../views/HomePage.vue'
 import NegotiationCreatePage from '../views/NegotiationCreatePage.vue'
-import NegotiationCreate from '../views/NegotiationCreate.vue'
 import NegotiationPage from '../views/NegotiationPage.vue'
 import FaqPage from '../views/FaqPage.vue'
 import NetworksPage from '../views/NetworksPage.vue'
@@ -64,33 +63,15 @@ const router = createRouter({
     },
     {
       path: '/requests/:requestId',
-      name: 'request',
+      name: 'request1',
       props: { isEditForm: false },
-      component: NegotiationCreatePage,
-    },
-    {
-      path: '/edit/requests/:requestId/:step(\\d+)',
-      name: 'editWithStep',
-      props: { isEditForm: true },
       component: NegotiationCreatePage,
     },
     {
       path: '/edit/requests/:requestId',
-      name: 'edit',
-      props: { isEditForm: true },
-      component: NegotiationCreatePage,
-    },
-    {
-      path: '/requests1/:requestId',
-      name: 'request1',
-      props: { isEditForm: false },
-      component: NegotiationCreate,
-    },
-    {
-      path: '/edit/requests1/:requestId',
       name: 'editWithStep1',
       props: { isEditForm: true },
-      component: NegotiationCreate,
+      component: NegotiationCreatePage,
     },
     {
       path: '/researcher',
